Imporve function documentation and apply minor fixes
- Removed "This function..." as its redundant.
- Switched to Google style docstrings for the Python
  bindings.
- There were a few places were I did a tiny refactor

+++ imgal_python/src/functions/statistics_functions.rs
@@ -7,15 +7,19 @@ use imgal::statistics;
 
 /// Compute the effective sample size (ESS) of a weighted sample set.
 ///
-/// This function computes the effective sample size (ESS) of a weighted sample
-/// set. Only the weights of the associated sample set are needed. The ESS is
-/// defined as:
+/// Computes the effective sample size (ESS) of a weighted sample set. Only the
+/// weights of the associated sample set are needed. The ESS is defined as:
 ///
+/// ```text
 /// ESS = (Σ wᵢ)² / Σ (wᵢ²)
+/// ```
 ///
-/// :param weights: A slice of non-negative weights where each element represents
-///     the weight of an associated sample.
-/// :return: The effective number of independent samples.
+/// Args:
+///     weights: A slice of non-negative weights where each element represents
+///         the weight of an associated sample.
+///
+/// Returns:
+///     The effective number of independent samples.
 #[pyfunction]
 #[pyo3(name = "effective_sample_size")]
 pub fn statistics_effective_sample_size(weights: Vec<f64>) -> f64 {
@@ -24,11 +28,14 @@ pub fn statistics_effective_sample_size(weights: Vec<f64>) -> f64 {
 
 /// Find the maximum value in an n-dimensional array.
 ///
-/// This function iterates through all elements of an n-dimensional array to
-/// determine the maximum value.
+/// Iterates through all elements of an n-dimensional array to determine the
+/// maximum value.
 ///
-/// :param data: The input n-dimensional array view.
-/// :return: The maximum value in the input data array.
+/// Args:
+///     data: The input n-dimensional array view.
+///
+/// Returns:
+///     The maximum value in the input data array.
 #[pyfunction]
 #[pyo3(name = "max")]
 pub fn statistics_max<'py>(data: Bound<'py, PyAny>) -> PyResult<f64> {
@@ -51,11 +58,14 @@ pub fn statistics_max<'py>(data: Bound<'py, PyAny>) -> PyResult<f64> {
 
 /// Find the minimum value in an n-dimensional array.
 ///
-/// This function iterates through all elements of an n-dimensional array to
-/// determine the minimum value.
+/// Iterates through all elements of an n-dimensional array to determine the
+/// minimum value.
 ///
-/// :param data: The input n-dimensional array view.
-/// :return: The minimum value in the input data array.
+/// Args:
+///     data: The input n-dimensional array view.
+///
+/// Returns:
+///     The minimum value in the input data array.
 #[pyfunction]
 #[pyo3(name = "min")]
 pub fn statistics_min<'py>(data: Bound<'py, PyAny>) -> PyResult<f64> {
@@ -78,13 +88,15 @@ pub fn statistics_min<'py>(data: Bound<'py, PyAny>) -> PyResult<f64> {
 
 /// Find the minimum and maximum values in an n-dimensional array.
 ///
-/// This function iterates through all elements of an n-dimensional array to
-/// determine the minimum and maximum values.
+/// Iterates through all elements of an n-dimensional array to determine the
+/// minimum and maximum values.
 ///
-/// :param data: The input n-dimensional array view.
-/// :return: A tuple containing the minimum and maximum values (_i.e._
-///     (min, max)) in the given array. If the array is empty a minimum and
-///     maximum value of 0 is returned in the tuple.
+/// Args:
+///     data: The input n-dimensional array view.
+///
+/// Returns:
+///     A tuple containing the minimum and maximum values (_i.e._ (min, max)) in
+///     the given array.
 #[pyfunction]
 #[pyo3(name = "min_max")]
 pub fn statistics_min_max<'py>(data: Bound<'py, PyAny>) -> PyResult<(f64, f64)> {
@@ -110,22 +122,43 @@ pub fn statistics_min_max<'py>(data: Bound<'py, PyAny>) -> PyResult<(f64, f64)> 
     }
 }
 
-/// Compute the linear percentile over an n-dimensonal array
+/// Compute the linear percentile over an n-dimensional array.
 ///
-/// This funtion computes the percentile of an entire array or percentiles along
-/// a given axis by creating 1-dimensional views along "axis".
+/// Calculates percentiles using linear interpolation between data points. The
+/// computation can be performed either on the entire array (flattened) or along
+/// a specified axis. The linear percentile is computed using:
 ///
-/// :param data: An n-dimensional image or array.
-/// :param p: The percentile value in the range of (0..100).
-/// :param axis: The axis to compute percentiles along. If "None", the input
-///     "data" is flattened and a single percentile value is returned.
-/// :param epsilon: The tolerance value used to decide the if the fractional
-///     index is an integer, default = 1e-12.
-/// :return: The linear percentile of the input data. If "axis" is
-///     "None", the result shape is "(1,)" and contains a single percentile
-///     value of the flattened input "data". If "axis" is a valid axis value,
-///     the result has the same shape as "data" with "axis" removed and contains
-///     the percentiles calculated along "axis".
+/// ```text
+/// h = (n - 1) × p
+/// j = ⌊h⌋
+/// γ = h - j
+/// percentile = (1 - γ) × v[j] + γ × v[j+1]
+/// ```
+///
+/// Where:
+/// - `n` is the array length.
+/// - `p` is the percentile in range `0` to `100`.
+/// - `v[j]` is the value at index `j`.
+/// - `⌊h⌋` is the floor function of `h`.
+///
+/// When `γ` is close to zero (within `epsilon`), the result is simply `v[j]`,
+/// avoiding unnecessary interpolation.
+///
+/// Args:
+///     data: An n-dimensional image or array.
+///     percentile: The percentile value in thae range `0.0` to `100.0`. Values
+///         out side this range will be clamped.
+///     axis: The axis to compute percentiles along. If `None`, the input `data`
+///         is flattened and a single percentile value is returned.
+///     epsilon: The tolerance value used to decide the if the fractional index
+///         is an integer. If `None`, then `epsilon = 1e-12`.
+///
+/// Returns:
+///     The linear percentile of the input data. If `axis` is `None`, the result
+///     shape is `(1,)` and contains a single percentile value of the flattened
+///     input `data`. If `axis` is a valid axis value, the result has the same
+///     shape as `data` with `axis` removed and contains the percentiles
+///     calculated along `axis`.
 #[pyfunction]
 #[pyo3(name = "linear_percentile")]
 #[pyo3(signature = (data, p, axis=None, epsilon=None))]
@@ -137,19 +170,19 @@ pub fn statistics_linear_percentile<'py>(
     epsilon: Option<f64>,
 ) -> PyResult<Bound<'py, PyArrayDyn<f64>>> {
     if let Ok(arr) = data.extract::<PyReadonlyArrayDyn<u8>>() {
-        statistics::linear_percentile(arr.as_array(), p as u8, axis, epsilon)
+        statistics::linear_percentile(arr.as_array(), p, axis, epsilon)
             .map(|output| output.into_pyarray(py))
             .map_err(map_imgal_error)
     } else if let Ok(arr) = data.extract::<PyReadonlyArrayDyn<u16>>() {
-        statistics::linear_percentile(arr.as_array(), p as u16, axis, epsilon)
+        statistics::linear_percentile(arr.as_array(), p, axis, epsilon)
             .map(|output| output.into_pyarray(py))
             .map_err(map_imgal_error)
     } else if let Ok(arr) = data.extract::<PyReadonlyArrayDyn<u64>>() {
-        statistics::linear_percentile(arr.as_array(), p as u64, axis, epsilon)
+        statistics::linear_percentile(arr.as_array(), p, axis, epsilon)
             .map(|output| output.into_pyarray(py))
             .map_err(map_imgal_error)
     } else if let Ok(arr) = data.extract::<PyReadonlyArrayDyn<f32>>() {
-        statistics::linear_percentile(arr.as_array(), p as f32, axis, epsilon)
+        statistics::linear_percentile(arr.as_array(), p, axis, epsilon)
             .map(|output| output.into_pyarray(py))
             .map_err(map_imgal_error)
     } else if let Ok(arr) = data.extract::<PyReadonlyArrayDyn<f64>>() {
@@ -163,10 +196,15 @@ pub fn statistics_linear_percentile<'py>(
     }
 }
 
-/// Compute the sum of a sequence of numbers.
+/// Compute the sum of the slice of numbers.
 ///
-/// :param data: The sequence of numbers.
-/// :return: The sum.
+/// Computes the sum of numbers in the input slice.
+///
+/// Args:
+///     data: A slice of numbers.
+///
+/// Returns:
+///     The sum.
 #[pyfunction]
 #[pyo3(name = "sum")]
 pub fn statistics_sum(data: Vec<f64>) -> f64 {
@@ -175,15 +213,17 @@ pub fn statistics_sum(data: Vec<f64>) -> f64 {
 
 /// Compute the weighted Kendall's Tau-b rank correlation coefficient.
 ///
-/// This function calculates a weighted Kendall's Tau-b rank correlation
-/// coefficient between two datasets. This implementation uses a weighted merge
-/// sort to count discordant pairs (inversions), and applies tie corrections for
-/// both variables to compute the final Tau-b coefficient. Here the weighted
+/// Calculates a weighted Kendall's Tau-b rank correlation coefficient between
+/// two datasets. This implementation uses a weighted merge sort to count
+/// discordant pairs (inversions), and applies tie corrections for both
+/// variables to compute the final Tau-b coefficient. Here the weighted
 /// observations contribute unequally to the final correlation coefficient.
 ///
 /// The weighted Kendall's Tau-b is calculated using:
 ///
-/// τ_b = (C - D) / √((n₀ - n₁)(n₀ - n₂))
+/// ```text
+/// τ_b = (C - D) / √[(n₀ - n₁)(n₀ - n₂)]
+/// ```
 ///
 /// Where:
 /// - `C` = number of weighted concordant pairs
@@ -192,14 +232,17 @@ pub fn statistics_sum(data: Vec<f64>) -> f64 {
 /// - `n₁` = weighted tie correction for first variable
 /// - `n₂` = weighted tie correction for second variable
 ///
-/// :param data_a: The first dataset for correlation analysis. Must be the same
-///     length as `data_b`.
-/// :param data_b: The second dataset for correlation analysis. Must be the same
-///     length as `data_a`.
-/// :param weights: The associated weights for each observation pait. Must be the
-///     same length as both input datasets.
-/// :return: The weighted Kendall's Tau-b correlation coefficient, ranging
-///     between -1.0 (negative correlation), 0.0 (no correlation) and 1.0
+/// Args:
+///     data_a: The first dataset for correlation analysis. Must be the same
+///         length as `data_b`.
+///     data_b: The second dataset for correlation analysis. Must be the same
+///         length as `data_a`.
+///     weights: The associated weights for each observation pait. Must be the
+///         same length as both input datasets.
+///
+/// Returns:
+///     The weighted Kendall's Tau-b correlation coefficient, ranging between
+///     `-1.0` (negative correlation), `0.0` (no correlation) and `1.0`
 ///     (positive correlation).
 #[pyfunction]
 #[pyo3(name = "weighted_kendall_tau_b")]
@@ -215,16 +258,22 @@ pub fn statistics_weighted_kendall_tau_b(
 
 /// Sort 1-dimensional arrays of values and their associated weights.
 ///
-/// This function performs a bottom up merge sort on the input 1-dimensional
-/// data array along with it's associated weights. Both the "data" and "weights"
-/// arrays are mutated during the sorting. The output of this function is a
-/// weighted inversion count.
+/// Performs a bottom up merge sort on the input 1-dimensional data array along
+/// with it's associated weights. Both the `data` and `weights` arrays are
+/// _mutated_ during the sorting. The output of this function is a weighted
+/// inversion count.
 ///
-/// :param data: A 1-dimensional array/slice of numbers of the same length as
-///    "weights".
-/// :param weights: A 1-dimensional array/slice of weights of the same length as
-///    "data".
-/// :return: The number of swaps needed to sort the input array.
+/// Args:
+///     data: A 1-dimensional array/slice of numbers of the same length as
+///         `weights`.
+///     weights: A 1-dimensional array/slice of weights of the same length as
+///         `data`.
+///
+/// Returns:
+///     The number of swaps needed to sort the input array.
+///
+/// Reference:
+///     <https://doi.org/10.1109/TIP.2019.2909194>
 #[pyfunction]
 #[pyo3(name = "weighted_merge_sort_mut")]
 pub fn statistics_weighted_merge_sort_mut<'py>(
